+++ Packages/com.mimylab.fukuroudon/Runtime/ActiveRelay/Scripts/ActiveRelayByInteract.cs
@@ -9,7 +9,6 @@ namespace MimyLab.FukuroUdon
     using UdonSharp;
     using UnityEngine;
     using VRC.SDKBase;
-    //using VRC.Udon;
 
     [Icon(ComponentIconPath.FukuroUdon)]
     [AddComponentMenu("Fukuro Udon/Active Relay/ActiveRelay by Interact")]
@@ -55,8 +54,10 @@ namespace MimyLab.FukuroUdon
 
         public override void Interact()
         {
-            DoAction();
-            Sync();
+            if (DoAction(Networking.LocalPlayer))
+            {
+                Sync();
+            }
         }
 
         private void Sync()