+++ Packages/com.mimylab.fukuroudon/Runtime/ActiveRelay/Scripts/ActiveRelayByPlayerCollision.cs
@@ -9,15 +9,18 @@ namespace MimyLab.FukuroUdon
     using UdonSharp;
     using UnityEngine;
     using VRC.SDKBase;
-    //using VRC.Udon;
+    using VRC.Udon.Common.Interfaces;
 
     [Icon(ComponentIconPath.FukuroUdon)]
     [AddComponentMenu("Fukuro Udon/Active Relay/ActiveRelay by Player Collision")]
+    [RequireComponent(typeof(Collider))]
     [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
     public class ActiveRelayByPlayerCollision : ActiveRelayBy
     {
         [SerializeField]
         private ActiveRelayPlayerEventType _eventType = default;
+        [SerializeField]
+        private NetworkEventTarget _acceptPlayerType = NetworkEventTarget.All;
 
         public override void OnPlayerCollisionEnter(VRCPlayerApi player)
         {
@@ -25,11 +28,7 @@ namespace MimyLab.FukuroUdon
             {
                 case ActiveRelayPlayerEventType.PlayerEnterAndExit:
                 case ActiveRelayPlayerEventType.PlayerEnter:
-                    DoAction();
-                    break;
-                case ActiveRelayPlayerEventType.LocalPlayerEnterAndExit:
-                case ActiveRelayPlayerEventType.LocalPlayerEnter:
-                    if (player.isLocal) { DoAction(); }
+                    if (CheckAccept(player)) { DoAction(player); }
                     break;
             }
         }
@@ -40,12 +39,25 @@ namespace MimyLab.FukuroUdon
             {
                 case ActiveRelayPlayerEventType.PlayerEnterAndExit:
                 case ActiveRelayPlayerEventType.PlayerExit:
-                    DoAction();
+                    if (CheckAccept(player)) { DoAction(player); }
                     break;
-                case ActiveRelayPlayerEventType.LocalPlayerEnterAndExit:
-                case ActiveRelayPlayerEventType.LocalPlayerExit:
-                    if (player.isLocal) { DoAction(); }
-                    break;
+            }
+        }
+
+        private bool CheckAccept(VRCPlayerApi player)
+        {
+            switch (_acceptPlayerType)
+            {
+                case NetworkEventTarget.All:
+                    return true;
+                case NetworkEventTarget.Owner:
+                    return player.IsOwner(this.gameObject);
+                case NetworkEventTarget.Others:
+                    return !player.isLocal;
+                case NetworkEventTarget.Self:
+                    return player.isLocal;
+                default:
+                    return false;
             }
         }
     }