+++ Packages/com.mimylab.fukuroudon/Runtime/SwivelChair2/Scripts/SwivelChair2.cs
@@ -9,7 +9,6 @@ namespace MimyLab.FukuroUdon
     using UdonSharp;
     using UnityEngine;
     using VRC.SDKBase;
-    //using VRC.Udon;
     using VRC.SDK3.Components;
 
     public enum SwivelChairPlayerPlatform